metatoolkit/requirements.txt
matplotlib
matplotlib_venn
networkx
numpy
pandas
pycircos
scipy
seaborn
shap
scikit-bio
scikit-learn
sklearn_som
statsmodels
umap
upsetplot
